fixing

+++ 0x15-file_io/0-read_textfile.c
@@ -13,7 +13,7 @@
 ssize_t read_textfile(const char *filename, size_t letters)
 {
 	ssize_t con1, n;
-	ssize_t fd;
+	int fd;
 	char buffer[letters];
 
 	if (filename == NULL)
